// src/_api/TaskManager/task-types.ts
export type LogLevel = 'silent' | 'error' | 'warn' | 'info' | 'debug' | 'trace'